Adding template
